--- a/Cargo.toml
+++ b/Cargo.toml
@@ -29,5 +29,10 @@ winapi-util = { version = "0.1.3", optional = true }
 encode_unicode = "0.3"
 
 [dev-dependencies]
+criterion = "0.3.5"
 proptest = "1.0.0"
 regex = "1.4.2"
+
+[[bench]]
+name = "ansi_parser"
+harness = false