--- a/src/ansi.rs
+++ b/src/ansi.rs
@@ -29,6 +29,7 @@ impl Default for State {
 
 impl State {
     fn is_final(&self) -> bool {
+        #[allow(clippy::match_like_matches_macro)]
         match self {
             Self::S3 | Self::S5 | Self::S6 | Self::S7 | Self::S8 | Self::S9 | Self::S11 => true,
             _ => false,
@@ -36,6 +37,7 @@ impl State {
     }
 
     fn is_trapped(&self) -> bool {
+        #[allow(clippy::match_like_matches_macro)]
         match self {
             Self::Trap => true,
             _ => false,