… .github/workflows/ci.yml
… tests/data/sample_zellij_session.log