# Motivation

`indicatif` recently removed their direct dependency on `regex`, but it still ends up getting pulled in through the `ansi-parsing` feature of this crate.

# Implementation

There's nothing too exciting here. The regex was switched out for a manually implemented dfa. The dfa is represented by `State` which is driven by `find_ansi_code_exclusive`. I went through a few iterations and the dfa implementation seemed to be the cleanest in the end.

There are also a myriad of tests to try and ensure that things match the old implementation. Notably there is a proptest checking off the old regex and a test that generates strings with bytes used in ansi codes up to 6 characters long in an attempt to ensure that everything matches the old implementation.

Also beyond that I tested things with `dfa_matches_regex_on_small_strings` with up to 11 byte strings using the limited character set and I ran another test that tried all valid 5 byte or small strings to make sure they matched the old regex `Matches`. This wasn't kept in the test suite since things would take far too long to give fast feedback.

# Impact

`regex` is a notably heavy dependency, so being able to remove it had some good improvements all around

## Compile Time

Clean `cargo build --release --features ansi-parsing`, timed with:

```cmd
$ hyperfine --warmup 1 --prepare 'cargo clean' 'cargo build --release --features ansi-parsing'
```

### Before

```text
Benchmark 1: cargo build --release --features ansi-parsing
  Time (mean ± σ):      5.395 s ±  0.061 s    [User: 28.892 s, System: 0.697 s]
  Range (min … max):    5.304 s …  5.531 s    10 runs
```

### After

```text
Benchmark 1: cargo build --release --features ansi-parsing
  Time (mean ± σ):      2.401 s ±  0.021 s    [User: 5.150 s, System: 0.405 s]
  Range (min … max):    2.374 s …  2.441 s    10 runs
```

## Size

I wasn't sure if there was an ideal way to measure this. I setup a crate that depended on just `console` and printed out the results of `AnsiCodeIterator` on the string `"hi"`. I also `strip`ped the executable for good measure since there was a large size reduction from stripping such a simple executable

Size results from `cargo build --release && strip target/release/testing`

*Before:* 653 KiB
*After:*  305 KiB

## Speed

I didn't directly test speed so this one is still unknown. I guess a good approach would be adding a benchmark to check the throughput for the new implementation iterating over something like `sample_zellij_session.log` and then backporting the benchmark to master to compare. From the small strings used in `dfa_matches_regex_on_small_string`, the new implementation was about 4x faster.