@LovecraftianHorror 2022-01-07
So I was curious about the speed changes, so I added in a basic benchmark that collects the results of `AnsiCodeIterator` to a vector along with timing `strip_ansi_codes` using the `zellij_session_data` that I already added as an asset used for a unit test. I ran the benchmark for these changes along with adding it to the current `master` with the one stipulation that I enabled all the default features since performance was pretty rough without the `perf` feature. With that being said here are the results I got on my laptop with a `11th Gen Intel i5-1135G7 (8) @ 4.200GHz` with it being plugged in to charge while running the benchmarks (there's a significant difference for me when it's on battery)

*Before:*

```text
ansi-parsing/parse      time:   [132.74 us 132.95 us 133.19 us]
                        thrpt:  [83.469 MiB/s 83.616 MiB/s 83.748 MiB/s]
ansi-parsing/strip      time:   [125.79 us 126.17 us 126.60 us]
                        thrpt:  [87.814 MiB/s 88.111 MiB/s 88.375 MiB/s]
```

*After:*

```text
ansi-parsing/parse      time:   [37.180 us 37.277 us 37.379 us]
                        thrpt:  [297.41 MiB/s 298.23 MiB/s 299.01 MiB/s]
ansi-parsing/strip      time:   [38.099 us 38.204 us 38.319 us]
                        thrpt:  [290.12 MiB/s 290.99 MiB/s 291.79 MiB/s]
```

So it looks like the changes did have a pretty significant impact even with a larger set of data that has quite a few ansi codes.