@mitsuhiko 2022-01-22
Fewer dependencies is always a huge plus in my book. Thank you so much for this!