@mitsuhiko 2022-01-22
@LovecraftianHorror unfortunately the PR appears to be using some features unavailable in the (quite old version of Rust) this is targeting. Now 1.40.0 is quite a bit older than what I think is sensible. I would probably be okay moving this up to 1.46.0 but not higher since that's something I'm targeting with insta and other libraries that use this still.