@LovecraftianHorror 2022-01-22
I should be able to make the changes compatible with 1.40.0 without too much effort. It is a pretty old version though like you said, so if you want to bump it then that works too!

I'm afk so I haven't checked, but I would assume this is compatible with 1.42.0 with the different pattern matching features used