@mitsuhiko 2022-01-22
I think moving this up to 1.42 or 1.46 is fine. Barely anything targets 1.40 these days.