@LovecraftianHorror 2022-01-23
- 6339a03a0ec32dd9818cda68fbe1a6758469e076 makes states addressed by the enum instead of the variants
- c548986b85ee82df383f90022176ee15708fedf2 removes all uses of or-patterns inside match destructuring (I forgot that stabilized so recently)
- f939e526738737c8d41afff50b48796dedf54688 switched out all uses of `matches!` for slightly more verbose `match`es

c548986b85ee82df383f90022176ee15708fedf2 brought the MSRV down to 1.42.0 and f939e526738737c8d41afff50b48796dedf54688 brought it down to 1.40.0 since it was such a simple change.