@LovecraftianHorror 2022-01-23
Just from a quick look, the older tests fail because both `criterion` and `proptest` don't build back on 1.40.0, even though `console` itself can still build back on 1.40.0

The clippy lints on the other hand are because of using `match` where it would be more appropriate to use `matches!` instead, but that change was specifically made to support an older version of rust.

I'm not sure how you want to approach this. Only testing on the most recent stable could help with the first note since libraries intending to be used as dev dependencies are more likely to use newer versions of rust. The second note could just be an `#[allow]` above the two cases where `matches!` was removed