@mitsuhiko 2022-01-23
I'm going to try to fix the CI setup. Unfortunately the version of rust we're targeting throws dev and non dev dependencies together which is why this is happening. I will probably going to move the tests up a few versions and only check against 1.40.